requirement.txt
absl-py==0.10.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
cachetools==4.1.1
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600699246375/work
chardet==3.0.4
click==7.1.2
colorama @ file:///tmp/build/80754af9/colorama_1603211150991/work
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-images-download==2.8.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598865816163/work
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1598353661586/work
Jinja2==2.11.2
joblib==0.17.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras-Preprocessing==1.1.2
keras-tuner==1.0.2
lazy-object-proxy==1.4.3
Markdown==3.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1601499549014/work
nltk==3.5
notebook @ file:///C:/ci/notebook_1601501643625/work
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.2
regex==2020.10.28
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-gpu==2.3.1
tensorflow-gpu-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
urllib3==1.25.10
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1603306325174/work